refactor(toolUse): consolidate shallow modules following design principles
Merged 4 shallow modules into 2 deeper modules:
- ToolUseSnapshotTypes.ts + ToolUseSnapshotCache.ts → ToolUseSessionManager.ts
- ToolUseFollowUpQueue.ts + ToolUseFollowUpCoordinator.ts → ToolUseFollowUp.ts

This reduces module sprawl and provides better abstractions by grouping
related functionality into cohesive units with clear responsibilities.
