+++ src/agent/toolUse/ToolUseFollowUp.ts
@@ -1,9 +1,29 @@
-// Local imports - logging
+/**
+ * Tool-use follow-up message handling.
+ *
+ * Provides:
+ * - Promise-based queue for follow-up messages in tool-use sessions
+ * - Static manager for queue instances indexed by stream ID
+ * - Coordination for routing follow-ups to active/resuming/pending sessions
+ */
+
+// Third-party imports
+import * as vscode from 'vscode';
+
+// Local imports
+import { getToolUseAgent } from '@agent/toolUse/ToolUseAgentRegistry';
 import type { StreamTabId } from '@agent/types/IdentifierTypes';
 import { AgentLogger } from '@logger/AgentLogger';
 
-const CHANNEL = 'ToolUseFollowUpQueue';
-const logger = new AgentLogger(CHANNEL);
+// Lazy imports to avoid circular dependency
+// ToolUseSessionManager and ToolUseSessionPersistence are imported at runtime
+import type { ToolUseSessionSnapshot } from './ToolUseSessionManager';
+
+const logger = new AgentLogger('ToolUseFollowUp');
+
+// ============================================================================
+// Follow-Up Queue (Instance)
+// ============================================================================
 
 /**
  * Promise-based queue for follow-up messages in a tool-use session.
@@ -103,6 +123,10 @@ export class FollowUpQueue {
   }
 }
 
+// ============================================================================
+// Follow-Up Queue Manager (Static)
+// ============================================================================
+
 /**
  * Static manager for follow-up queues indexed by stream ID.
  */
@@ -175,4 +199,76 @@ export class ToolUseFollowUpQueue {
   static get(streamId: StreamTabId): FollowUpQueue | undefined {
     return this.queues.get(streamId);
   }
+}
+
+// ============================================================================
+// Follow-Up Coordination
+// ============================================================================
+
+/**
+ * Send a follow-up message to a tool-use session.
+ *
+ * Routes the message based on session state:
+ * 1. Active agent: direct append
+ * 2. Resuming session: queue for later
+ * 3. Pending snapshot: lazy resume with follow-up
+ * 4. No session: show warning
+ */
+export async function sendFollowUp(
+  streamId: StreamTabId,
+  text: string,
+): Promise<void> {
+  // Try active agent first
+  const agent = getToolUseAgent(streamId);
+  if (agent) {
+    try {
+      agent.appendFollowUp(text);
+    } catch (error) {
+      logger.error('Failed to send follow-up to active agent.', {
+        data: error,
+      });
+      await vscode.window.showErrorMessage(
+        `Failed to send follow-up: ${(error as Error).message}`,
+      );
+    }
+    return;
+  }
+
+  // Queue if session is resuming
+  if (ToolUseFollowUpQueue.isResuming(streamId)) {
+    if (ToolUseFollowUpQueue.enqueue(streamId, text)) {
+      logger.debug(`Queued follow-up while stream ${streamId} is resuming.`);
+      return;
+    }
+  }
+
+  // Lazy resume from snapshot if available
+  const { ToolUseSessionManager } = await import('./ToolUseSessionManager');
+  const { ToolUseSessionPersistence } =
+    await import('./ToolUseSessionPersistence');
+
+  const pendingSnapshot = ToolUseSessionManager.getByStream(streamId);
+  if (pendingSnapshot) {
+    logger.debug(`Resuming agent lazily for stream ${streamId}.`);
+    await ToolUseSessionPersistence.resumeFromSnapshot(pendingSnapshot, text);
+    return;
+  }
+
+  // No session found
+  logger.debug(`No active session found for follow-up on stream ${streamId}.`);
+  void vscode.window.showWarningMessage(
+    'No active tool-use session found for this follow-up.',
+  );
+}
+
+/**
+ * Resume an agent from a persisted snapshot.
+ */
+export async function resumeFromSnapshot(
+  snapshot: ToolUseSessionSnapshot,
+  followUp?: string,
+): Promise<{ success: boolean; lostFollowUps?: number }> {
+  const { ToolUseSessionPersistence } =
+    await import('./ToolUseSessionPersistence');
+  return ToolUseSessionPersistence.resumeFromSnapshot(snapshot, followUp);
 }
