+++ src/agent/toolUse/index.ts
@@ -12,9 +12,7 @@ export { BaseToolUseAgent } from '@agent/implementations/BaseToolUseAgent';
 export * from './ToolUseAgentRegistry';
 export * from './ToolUseSessionPersistence';
 export * from './ToolUseSessionLifecycle';
-export * from './ToolUseFollowUpCoordinator';
-export * from './ToolUseFollowUpQueue';
-export * from './ToolUseSnapshotCache';
+export * from './ToolUseFollowUp';
+export * from './ToolUseSessionManager';
 export * from './ToolUseSnapshotStore';
-export * from './ToolUseSnapshotTypes';
 export * from './ToolFileInteractionContext';
